# openapi.yaml
openapi: 3.0.3
info:
  title: HR Agent AI — Backend API
  version: "1.1.0"
  description: |
    Backend для AI-ассистента «Техна» по HR-вопросам компании 1221 Системс.
    Полностью локальная архитектура: Node.js/Express + PostgreSQL + Ollama
    (qwen2.5:32b для генерации, bge-m3 для эмбеддингов) + локальный
    RAG-индекс по ЛНА (ПВТР, положения, инструкции).

    ## Аутентификация и роли

    Все защищённые эндпоинты принимают JWT в заголовке `Authorization: Bearer <token>`.
    JWT выдаётся `POST /api/auth/login` и содержит поле `role`:
      * `employee` — обычный сотрудник;
      * `admin` — администратор.

    Тестовые учётки:
      * `admin / admin` — администратор;
      * `12345 / password123` (Потапов), `67890 / password123` (Иванова),
        `11111 / password123` (Петров) — обычные сотрудники.

    ## Матрица доступа

    | Эндпоинт                              | Аноним | Сотрудник                | Админ |
    | ------------------------------------- | ------ | ------------------------ | ----- |
    | `GET  /health`                        |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/register`             |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/login`                |  ✅    |  ✅                      |  ✅    |
    | `POST /api/auth/verify`               |  ❌    |  ✅                      |  ✅    |
    | `POST /api/employee/auth`             |  ✅    |  ✅                      |  ✅    |
...

    Возможности:
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/chat/message:
    post:
      tags: [Chat]
      summary: Задать вопрос ассистенту (синхронный ответ)
      description: |
        **Требует JWT.**

        Извлекает релевантные фрагменты из RAG, подмешивает данные сотрудника
        и историю диалога, генерирует ответ на qwen2.5:32b.

        * **Сотрудник:** `employeeId` всегда берётся из токена, поле `employeeId`
          в body **игнорируется**. Чат ведётся от имени самого пользователя.
        * **Админ:** может передать `employeeId` в body и задавать вопросы
          «от имени» этого сотрудника (история и контекст подтянутся для него).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              vacation:
                summary: «Сколько у меня дней отпуска?» (любой авторизованный)
                value:
                  message: "Сколько у меня осталось дней отпуска?"
              admin_for_user:
                summary: Админ задаёт вопрос от имени сотрудника 12345
                value:
                  message: "Какой у меня график отпусков?"
                  employeeId: "12345"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /api/chat/stream:
    post:
      tags: [Chat]
      summary: Тот же ответ, но потоком через Server-Sent Events
      description: |
        **Требует JWT.** Те же правила по `employeeId`, что и `/api/chat/message`:
        обычный юзер пишет от себя, админ может указать `employeeId` в body.

        Возвращает поток `text/event-stream`. Тело каждого события — JSON одного из видов:
          * `{"type":"context","source":"...","hits":[...]}` — однократно в начале
          * `{"type":"token","delta":"..."}` — токены ответа
          * `{"type":"done"}` — конец стрима
          * `{"type":"error","error":"..."}` — ошибка
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: SSE-поток
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type":"context","source":"ПВТР, п. 7.6","hits":[]}

                  data: {"type":"token","delta":"У вас "}

                  data: {"type":"token","delta":"14 дней."}

                  data: {"type":"done"}
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/chat/history/{employeeId}:
    get:
      tags: [Chat]
      summary: История чата сотрудника
      description: |
        **Требует JWT.**
        * Сотрудник: только свой `employeeId`, иначе 403.
        * Админ: любой `employeeId`.
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatHistoryItem"
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/documents:
    get:
      tags: [Documents]
      summary: Список документов
      description: "**Требует JWT** (любая роль)."
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items: { $ref: "#/components/schemas/Document" }
        "401": { $ref: "#/components/responses/Error" }
  /api/documents/{id}:
    get:
      tags: [Documents]
      summary: Документ по id
      description: "**Требует JWT** (любая роль)."
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Documents]
      summary: Удалить документ (вместе с чанками RAG) — admin only
      description: "**Только admin.** Обычный сотрудник получит 403."
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  removedChunks: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/documents/upload:
    post:
      tags: [Documents]
      summary: Загрузка нового документа (PDF/DOCX/TXT/MD) — admin only
      description: |
        **Только admin.** Загружает файл, парсит, индексирует в RAG и сохраняет
        метаданные в БД. Обычный сотрудник получит 403.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:    { type: string, example: "Положение о премировании" }
                category: { type: string, example: "Заработная плата" }
                type:     { type: string, example: "ЛНА" }
      responses:
        "201":
          description: Загружен и проиндексирован
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Document"
                  - type: object
                    properties:
                      chunkCount: { type: integer, example: 17 }
                      message:    { type: string }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/documents/meta/categories:
    get:
      tags: [Documents]
      summary: Список категорий
      description: "**Требует JWT** (любая роль)."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EmployeeId:
      in: path
      name: id
      required: true
      schema: { type: string }
      example: "12345"
      description: |
        Табельный номер. Обычный сотрудник может указывать только собственный
        `employeeId` (иначе 403). Админ — любой существующий.
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:    { type: string, example: "Сколько у меня осталось дней отпуска?" }
        employeeId:
          type: string
          example: "12345"
          description: |
            Игнорируется для обычного сотрудника (`employeeId` всегда берётся из JWT).
            Админ может указать чужой `employeeId`, чтобы вести чат от имени сотрудника.
    RagHit:
      type: object
      properties:
        score:    { type: number, example: 0.66 }
        doc:      { type: string, example: "Правила внутреннего трудового распорядка" }
        section:  { type: string, example: "п. 7.6" }
    ChatResponse:
      type: object
      properties:
        response:  { type: string }
        source:    { type: string, nullable: true }
        hits:
          type: array
          items: { $ref: "#/components/schemas/RagHit" }
        timestamp: { type: string, format: date-time }
    ChatHistoryItem:
      type: object
      properties:
        id:        { type: string }
        role:      { type: string, enum: [user, assistant] }
        content:   { type: string }
        timestamp: { type: string, format: date-time }
    Document:
      type: object
      properties:
        id:          { type: string }
        title:       { type: string }
        type:        { type: string }
        category:    { type: string }
        uploadDate:  { type: string, nullable: true }
        fileUrl:     { type: string }
        chunkCount:  { type: integer, nullable: true }
        indexed:     { type: boolean, nullable: true }